#2442 Refactor common-canvas.ts for better structure
